requirements.txt
selenium
beautifulsoup4
pyspark
minio
psycopg2-binary
pyyaml
loguru
tqdm
